refactor: update database schema structure with dynamic table loading
