+++ changes/2025-03-15.md
@@ -1,3 +1,15 @@
+## Implemented Audience Tracking for Waitlist
+
+- Added audience tracking functionality to capture referral sources and UTM parameters
+- Created database schema for storing audience tracking data with proper relations to waitlist entries
+- Implemented migration file for the new audience tracking table
+- Developed utility functions for extracting tracking data from request objects
+- Created a service layer for managing audience tracking data operations
+- Updated waitlist API to capture and store tracking information
+- Added comprehensive documentation for the audience tracking feature
+- Ensured type safety throughout the implementation
+- Validated changes with successful build and type checking
+
 ## UI Enhancements
 
 - Improved UI components for better user experience
