+++ packages/database/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib';
-export * from './schema';
+export * from './tables';
+export * from './schemas';
 export * from './services';
