--- packages/database/src/schema/index.ts
@@ -1,2 +1,0 @@
-export * from './waitlist.schema';
-export * from './audience-tracking.schema';
