+++ packages/database/src/schemas/audience-tracking/insert.type.ts
@@ -1,0 +1,7 @@
+import { type InferInsertModel } from 'drizzle-orm';
+import { audienceTrackingTable } from '../../tables/audience-tracking.table';
+
+// Type definition for inserting audience tracking data
+export type AudienceTrackingInsert = InferInsertModel<
+    typeof audienceTrackingTable
+>;
