+++ packages/database/src/schemas/audience-tracking/query-params.type.ts
@@ -1,0 +1,11 @@
+// Query parameters for filtering audience tracking data
+export type AudienceTrackingQueryParams = {
+    startDate?: Date;
+    endDate?: Date;
+    utm_source?: string;
+    utm_medium?: string;
+    utm_campaign?: string;
+    location_country?: string;
+    limit?: number;
+    offset?: number;
+};
