+++ packages/database/src/schemas/audience-tracking/result.type.ts
@@ -1,0 +1,8 @@
+import { AudienceTrackingSelect } from './select.type';
+
+// Result type for audience tracking service methods
+export type AudienceTrackingResult = {
+    success: boolean;
+    data?: AudienceTrackingSelect;
+    error?: string;
+};
