+++ packages/database/src/schemas/audience-tracking/select.type.ts
@@ -1,0 +1,7 @@
+import { type InferSelectModel } from 'drizzle-orm';
+import { audienceTrackingTable } from '../../tables/audience-tracking.table';
+
+// Type definition for selecting audience tracking data
+export type AudienceTrackingSelect = InferSelectModel<
+    typeof audienceTrackingTable
+>;
