+++ packages/database/src/schemas/index.ts
@@ -1,0 +1,2 @@
+export * from './audience-tracking';
+export * from './waitlist';
