+++ packages/database/src/schemas/waitlist/entries-result.type.ts
@@ -1,0 +1,7 @@
+import { WaitlistSelect } from './select.type';
+
+// Result type for listing waitlist entries
+export type WaitlistEntriesResult = {
+    entries: WaitlistSelect[];
+    count: number;
+};
