+++ packages/database/src/schemas/waitlist/form-data.type.ts
@@ -1,0 +1,6 @@
+// Form data type for waitlist submissions
+export type WaitlistFormData = {
+    name: string;
+    email: string;
+    company?: string;
+};
