+++ packages/database/src/schemas/waitlist/index.ts
@@ -1,0 +1,6 @@
+export * from './schema';
+export * from './select.type';
+export * from './insert.type';
+export * from './form-data.type';
+export * from './result.type';
+export * from './entries-result.type';
