+++ packages/database/src/schemas/waitlist/insert.type.ts
@@ -1,0 +1,5 @@
+import { type InferInsertModel } from 'drizzle-orm';
+import { waitlistTable } from '../../tables/waitlist.table';
+
+// TypeScript type using Drizzle's type inference
+export type WaitlistInsert = InferInsertModel<typeof waitlistTable>;
