+++ packages/database/src/schemas/waitlist/result.type.ts
@@ -1,0 +1,8 @@
+import { WaitlistSelect } from './select.type';
+
+// Result type for waitlist service methods
+export type WaitlistResult = {
+    success: boolean;
+    data?: WaitlistSelect;
+    error?: string;
+};
