+++ packages/database/src/tables/index.ts
@@ -1,0 +1,2 @@
+export * from './waitlist.table';
+export * from './audience-tracking.table';
